src/pyproject.toml
[project]
name = "ai-honeypot"
version = "1.0.0"
description = "AI-Driven Honeypot System"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.0",

    # ML (CPU only)
    "transformers>=4.36.0",
    "torch",
    # torch installed as CPU-only via Dockerfile pip command

    # DB + backend
    "sqlalchemy==2.0.32",
    "aiosqlite==0.20.0",
    "asyncpg>=0.31.0",

    # utils
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "pydantic==2.7.4",
    "pydantic-settings>=2.13.1",
    "httpx==0.27.0",
    "loguru==0.7.2",
    "PyJWT==2.8.0",
    "redis==5.0.8",
    "supabase>=2.0.0",

    # migrations
    "alembic>=1.18.4",
    "greenlet>=3.4.0",
    "slowapi==0.1.9",
    "prometheus-fastapi-instrumentator==7.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app", "src/core"]

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
